+++ packages/widgets/src/components/SegmentedPieChart/index.ts
@@ -1,0 +1,1 @@
+export * from './SegmentedPieChart';
